+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.x.x] - 2021-xx-xx
 * Updated `cosmic_gps` to support xarray Datasets
+* Added altitude binning profile support for all datasets with altitude
+  as a variable.
 * Added `download` to general methods
 * Update links in documentation